+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Refactor boilerplate and every game as an independent project
 * Add new `experimental` project `raycasting` *[Author: digitsensitive (2020-03-05)]*
 * Add new `experimental` project `untextured-raycaster` *[Author: digitsensitive (2020-03-06)]*
+* Add new `experimental` project `game-of-life` *[Author: digitsensitive (2020-03-07)]*
 
 ### Updates